+++ wperf/parsers.cpp
@@ -174,7 +174,7 @@ void parse_events_extra(std::wstring events_str, std::map<enum evt_class, std::v
     $ perf record -e arm_spe/branch_filter=1,jitter=1/ -- workload
     $ perf record -e spe/branch_filter=1,jitter=1/ -- workload
 */
-bool parse_events_str_for_feat_spe(std::wstring events_str, std::map<std::wstring,uint32_t>& flags)
+bool parse_events_str_for_feat_spe(std::wstring events_str, std::map<std::wstring,uint64_t>& flags)
 {
     // Detect arm_spe_0/*/      - where '*'
     if (events_str.size() >= std::wstring(L"arm_spe_0//").size()
